% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\alias{extract.Table}
\alias{extract.THsegments}
\title{extracting clusters}
\usage{
extract(segimg, ..., cluster = NULL, org_image = NULL, b = NULL)

\method{extract}{Table}(segimg, cluster = 1, org_image = NULL,
  b = NULL, ...)

\method{extract}{THsegments}(segimg, ..., cluster = 1)
}
\arguments{
\item{segimg}{an object of class Table or THsegments or grayimage}

\item{...}{graphical parameters to imageplot}

\item{cluster}{cluster number to be extracted}

\item{org_image}{original image to be segmented}

\item{b}{number of bands in the original image}
}
\description{
Documentation for the extract method of single cluster/class from segmented images
}
\examples{
\dontrun{
library(stars)
tif <- read_stars(system.file("extdata/image.tif", package="ImageSegmentation"))
p <- seg_kmeans(tif,3,5)
extract(p,2,tif,3)

e <- RGBtoGRAY(tif)
s <- segmentation_threshold(e)
extract(s, cluster=3)

}


}
