% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageSegmentation.R
\docType{package}
\name{ImageSegmentation}
\alias{ImageSegmentation}
\alias{ImageSegmentation-package}
\title{The ImageSegmentation Package}
\description{
This package contains functions to segment images.For a complete list of functions, use library (help = "ImageSegmentation").
}
\author{
Chamodi Boyagoda \ email{: chamodib88@gamil.com}
}
