% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgb2gray.R
\name{RGBtoGRAY}
\alias{RGBtoGRAY}
\title{converting RGB image to gray image}
\usage{
RGBtoGRAY(m, n = 3, Red = 1, Green = 2, Blue = 3)
}
\arguments{
\item{m}{Input image}

\item{n}{Number of bands in the image (default is 3)}

\item{Red}{Band number of red (default is 1)}

\item{Green}{Band number of green (default is 2)}

\item{Blue}{Band number of blue (default is 3)}
}
\value{
an object of class "grayimage"
}
\description{
Documentation for the RGBtoGRAY function
}
\examples{
\dontrun{
library(stars)
imgRGB <- read_stars(system.file("extdata/image.tif", package="ImageSegmentation"))
RGBtoGRAY(imgRGB,3,1,2,3)
}

}
