% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SegKmeans.R
\name{seg_kmeans}
\alias{seg_kmeans}
\title{Perform segmentation using kmeans algorithm on an image}
\usage{
seg_kmeans(im, b = 3, k = 5)
}
\arguments{
\item{im}{Input image}

\item{b}{Number of bands in the image (default is 3)}

\item{k}{Number of clusters (default is 5)}
}
\description{
Documentation for the seg_kmeans function
}
\examples{
\dontrun{
library(stars)
tif <- read_stars(system.file("extdata/image.tif", package="ImageSegmentation"))
j <- seg_kmeans(tif)
}

}
