% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SegThreshold.R
\name{segmentation_threshold}
\alias{segmentation_threshold}
\title{Perform segmentation on a gray image}
\usage{
segmentation_threshold(p, val1 = mean, val2 = 0.5 * mean, val3 = 0.25
  * mean)
}
\arguments{
\item{p}{Input an object of class "grayimage"}

\item{val1}{Threshold value 1 (default is mean of all pixel values)}

\item{val2}{Threshold value 2 (default is half of mean)}

\item{val3}{Threshold value 3 (default is one quarter of mean)}
}
\value{
returns an object of class "THsegments"
}
\description{
Documentation for the segmentation_threshold function
}
\examples{
\dontrun{
library(stars)
pic <- read_stars(system.file("extdata/image.tif", package="ImageSegmentation"))
g <- RGBtoGRAY(pic,3,1,2,3)
segmentation_threshold(g)
}

}
